Deleted duplicate files
