--- abdi-README.md
@@ -1,14 +1,0 @@
-# autocomplete
-easily find words in the English dictionary
-
-## Features
-
-### Autocomplete module
-+ [x]  import a list of words into an array
-+ called words  
-+ [x] search through a list of words for a string and return list of suggestions  
-+ [x] record the string that was searched for (for stats and graphs!)
-
-### Autocomplete HTTP Server (API)
-+ [ ] serve an html page with a search box
-+ [ ] expose the findWords method as an API endpoint
