--- abdi-package.json
@@ -1,34 +1,0 @@
-{
-  "name": "autocomplete-fac",
-  "version": "1.0.2",
-  "description": "find words in English dictionary with auto suggestion",
-  "main": "index.js",
-  "directories": {
-    "test": "test"
-  },
-  "scripts": {
-    "test": "node test/test.js",
-    "coverage": "istanbul cover test/test.js",
-    "nodemon": "./node_modules/.bin/nodemon server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/minaorangina/autocomplete.git"
-  },
-  "keywords": [
-    "autocomplete",
-    "english",
-    "dictionary",
-    "suggestions"
-  ],
-  "author": "minaorangina",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/minaorangina/autocomplete/issues"
-  },
-  "homepage": "https://github.com/minaorangina/autocomplete",
-  "devDependencies": {
-    "istanbul": "^0.3.14",
-    "nodemon": "^1.3.7"
-  }
-}
