+++ index.html
@@ -11,16 +11,15 @@
     </div>
 
     <style>
-
 body {
   min-height: 100%;
   height: 100%
 }
 #main {
-  height: 30%;
+  height: 50%;
 }
 #quint {
-  height: 70%;
+  height: 50%;
   bottom: 0;
   position: fixed;
 }
@@ -31,25 +30,24 @@ body {
     <div id="qunit-fixture"></div>
     <script src="https://code.jquery.com/qunit/qunit-1.18.0.js"></script>
 
-    <script>
+	<script>
+// 	console.log('Test');
+// test('Search for words beginning with awes', function(assert){
+//   var done = assert.async();
+//   $.get('/find/awes', function handler(data) {
+//     console.log(data);
+//     var words = data.split(',');
+//     assert.equal(words.length, 4, 'number of suggestions for awes is 4');
+//   });
+// });
+//     </script>
 
-test('Search for words beginning with awes', function(assert){
-  var done = assert.async();
-  $.get('/find/awes', function handler(data) {
-    console.log(data);
-    var words = data.split(',');
-    assert.equal(words.length, 4, 'number of suggestions for awes is 4');
-  });
-});
 
-    </script>
-
-
-<script>
+ <script>
   $('#search').keyup(function(e) {
     var word = $('#search').val();
-    if (word.length > 2) {
-      $.get('/find/'+word, function handler(data) {
+    if (word.length > 2) {//if word length greater than 2 make ajax call
+      $.get('/find/'+ word, function handler(data) {
         console.log(data);
         var words = data.split(',');
         var results = '';
@@ -60,7 +58,7 @@ test('Search for words beginning with awes', function(assert){
       });
     }
   })
-</script>
+// </script>
 
   </body>
 </html>
