+++ index.js
@@ -1,39 +1,70 @@
 var fs = require('fs');
+var http = require('http'); 
 var ac = {};
 
-ac.import = function (callback) {
-  if (!callback || typeof callback !== 'function') {
+ac.import = function(callback){ // saving import fct as a property of ac{}
+  if(!callback || typeof callback !== 'function'){
     return new Error('callback argument MUST be a function');
   }
   var filename = __dirname + '/words.txt';
-  fs.readFile(filename, 'utf8', function (err, data) {
-    ac.words = data.split('\n');
-    return callback(err, ac.words);
+  fs.readFile(filename, 'utf8', function(err, data){ // fs.readFile(myFile, ['utf8'], callback)
+    ac.words = data.split('\n'); // stores array of words in ac{}
+    return callback(err, ac.words); // call callback function on ac.words
   });
-};
+}
 
-ac.stats = function (word, callback) {
-  if (!ac.searches) {
+ac.stats = function(word, callback){
+  if (!ac.searches){
     ac.searches = {};
   }
-  if (!ac.searches[word]) {
+  if (!ac.searches[word]){
     ac.searches[word] = [];
   }
   ac.searches[word].push(new Date().getTime());
   callback(null, ac.searches);
-};
+}
 
-ac.findWord = function (word, callback) {
-  // who wants to volunteer to implement the method?
+ac.findWord = function (word, callback){
+  //var found = ac.words[ac.words.indexOf(word)];
   var found = [];
-  for (var i = 0; i < ac.words.length; i++) {
-    if (ac.words[i].search(word) === 0) {
+  for (var i = 0; i < ac.words.length; i++){
+    if (ac.words[i].search(word) === 0){
       found.push(ac.words[i]);
     }
   }
-  return callback(null, found);
-};
+  return callback(null,found);
+}
 
+// record timestamp at which a search was made
 
+//define method on ac module
+ac.define = function (word, callback){
+	var wordDef
+	var body = ''; 
+	var url = 'http://api.wordnik.com/v4/word.json/'+ word +'/definitions?limit=1&includeRelated=true&useCanonical=false&includeTags=false&api_key=a2a73e7b926c924fad7001ca3111acd55af2ffabf50eb4ae5';
 
-module.exports = ac;
+	var request = http.get(url, function(response) {
+		
+		//Read the data
+        response.on('data', function (chunk) {
+            body += chunk;
+
+        });
+
+        response.on('end', function () {
+            if(response.statusCode === 200) {
+               
+                wordDef = JSON.parse(body);
+                //console.log(body);
+                return callback(null, wordDef);
+            }
+        });
+
+		
+	});
+
+	 
+	
+}
+
+module.exports = ac;// exporting ac object to use elsewhere
