+++ server.js
@@ -1,6 +1,8 @@
 var http = require('http');
 var port = process.env.PORT || 3000;
 var ac = require('./index.js');
+
+//load in the words from the file
 ac.import(function(err, count) {
   console.log("imported a bunch of words! >> ", count.length);
 });
@@ -15,6 +17,7 @@ http.createServer(function handler(request, response) {
     response.writeHead(200, {"Content-Type": "text/html"});
     response.end(index.toString());
   }
+
   if (url.indexOf('/find/') > -1) {
     // locahost:3000/find/word
     var word = url.split('/')[2];
@@ -25,6 +28,19 @@ http.createServer(function handler(request, response) {
     });
     // response.end('word: ', word);
   }
+
+  if(url.indexOf('/test/')){
+  	ac.define('shirt', function(err, definition){
+  		 
+  		response.end(definition[0].text); 
+  		console.log('definition: '+definition[0].text);  
+  	}); 
+  		
+  		
+	// console.log('your requested this url: '+ url); 
+	// console.log('------------------------------')
+  }
+
   else {
     response.end('hello Dan!');
   }
@@ -32,3 +48,4 @@ http.createServer(function handler(request, response) {
 }).listen(port);
 
 console.log('node http server listening on http://localhost:' + port);
+console.log('Test');
